// packages/service/types/AuctionInitiator.ts
export interface AuctionInitiatorInterface extends Interface {
  getFunction(
    nameOrSignature: "createAuction" | "getAllAuctions" | "getAuctionsByCreator"
  ): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "AuctionCreated"): EventFragment;

  encodeFunctionData(
    functionFragment: "createAuction",
    values: [string, string, BigNumberish, BigNumberish, BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getAllAuctions",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "getAuctionsByCreator",
    values: [AddressLike]
  ): string;

  decodeFunctionResult(
    functionFragment: "createAuction",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getAllAuctions",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getAuctionsByCreator",
    data: BytesLike
  ): Result;
}

export namespace AuctionCreatedEvent {
  export type InputTuple = [creator: AddressLike, auction: AddressLike];
  export type OutputTuple = [creator: string, auction: string];
  export interface OutputObject {
    creator: string;
    auction: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface AuctionInitiator extends BaseContract {
  connect(runner?: ContractRunner | null): AuctionInitiator;
  waitForDeployment(): Promise<this>;

  interface: AuctionInitiatorInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  createAuction: TypedContractMethod<
    [
      tokenName: string,
      tokenSymbol: string,
      totalSupply: BigNumberish,
      startPrice: BigNumberish,
      reservedPrice: BigNumberish
    ],
    [string],
    "nonpayable"
  >;

  getAllAuctions: TypedContractMethod<[], [string[]], "view">;

  getAuctionsByCreator: TypedContractMethod<
    [creator: AddressLike],
    [string[]],
    "view"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "createAuction"
  ): TypedContractMethod<
    [
      tokenName: string,
      tokenSymbol: string,
      totalSupply: BigNumberish,
      startPrice: BigNumberish,
      reservedPrice: BigNumberish
    ],
    [string],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "getAllAuctions"
  ): TypedContractMethod<[], [string[]], "view">;
  getFunction(
    nameOrSignature: "getAuctionsByCreator"
  ): TypedContractMethod<[creator: AddressLike], [string[]], "view">;

  getEvent(
    key: "AuctionCreated"
  ): TypedContractEvent<
    AuctionCreatedEvent.InputTuple,
    AuctionCreatedEvent.OutputTuple,
    AuctionCreatedEvent.OutputObject
  >;

  filters: {
    "AuctionCreated(address,address)": TypedContractEvent<
      AuctionCreatedEvent.InputTuple,
      AuctionCreatedEvent.OutputTuple,
      AuctionCreatedEvent.OutputObject
    >;
    AuctionCreated: TypedContractEvent<
      AuctionCreatedEvent.InputTuple,
      AuctionCreatedEvent.OutputTuple,
      AuctionCreatedEvent.OutputObject
    >;
  };
}